requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
python-dotenv
openpyxl
google-generativeai
tabulate